requirements.txt
docutils
pygments
pypiwin32
kivy_deps.sdl2==0.1.*
kivy_deps.glew==0.1.*
kivy_deps.gstreamer==0.1.*
kivy_deps.angle==0.1.*
kivy==1.11.1
numpy
opencv
Pillow
pandas
wget
lxml
markdown
